+++ docs/change_log.md
@@ -84,6 +84,7 @@ There is one key exception to the rules above -- and that is with `MAJOR`=0 rele
 - add `simmate engine start-schedules` which let's you configure periodic tasks for individual apps (e.g. check a table for updates every 5 minutes). Includes error handling and email alerts. (Note: this a quick alternative to full Prefect system)
 - add warren_lab app with Warren Lab preferred VASP settings
 - add badelf app with class oriented tools for performing BadELF analyses
+- add basic quantum espresso support
 
 **Refactors**